requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.1.0
autobahn==24.4.2
Automat==22.10.0
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
django-braces==1.15.0
django-debug-toolbar==4.3.0
django-embed-video==1.4.9
django-redisboard==8.4.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
pillow==10.3.0
psycopg==3.1.18
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pymemcache==4.0.0
pyOpenSSL==24.2.1
python-decouple==3.8
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
rpds-py==0.19.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uWSGI==2.0.25.1
zope.interface==7.0.1
